backend/mysql-mcp-server/requirements.txt
fastmcp>=0.1.0
pymysql>=1.1.0
python-dotenv>=1.0.0
cryptography>=42.0.0
